requirements.txt
pyyaml==3.12
unirest==1.1.7
oauth2client == 3.0.0
google-api-python-client == 1.6.5
CLick == 6.7
python-ldap == 2.4.28